Add backtrace to exceptions treated as fatal errors w/ test, and move partial backtrace reconstruction to FrontController so different safemodes can receive it.
